# api/openapi.yaml
openapi: 3.1.0
info:
  title: Trident REST API
  description: |
    Real-time indexing and event querying for Soroban contract events.

    ## Request validation

    Every endpoint validates its parameters before doing any work and answers a
    malformed request with `400` and the `ErrorResponse` envelope, where
    `error.code` is `INVALID_ARGUMENT` and `error.message` names the field that
    failed:

    ```json
    {"error": {"code": "INVALID_ARGUMENT", "message": "limit must be an integer between 1 and 200"}}
    ```

    Rules shared across endpoints:

    - `contractId` — Stellar contract strkey: `C` followed by 55 uppercase
      base32 characters.
    - path ids — canonical UUID v4.
    - ledger ranges — non-negative integers; the upper bound must be greater
      than or equal to the lower bound.
    - `limit` — an integer within the range documented for that endpoint.
    - `cursor` — an opaque cursor from a previous response; anything else is
      rejected rather than silently ignored.
    - `network` — `testnet` or `mainnet`.
    - Unknown query parameters are rejected, so a typo such as `?limitt=5`
      fails loudly instead of silently changing the page size.
  version: 1.0.0
  contact:
    name: Trident Team
  license:
    name: Apache 2.0
servers:
  - url: https://api.trident.telocel.com
    description: Production
  - url: http://localhost:3000
    description: Development
security:
  - ApiKeyAuth: []
paths:
...
  /v1/api-keys:
    post:
      summary: Create an API key
      description: Generate a new API key for authentication
      operationId: createApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  description: Human-readable label for this key
                network:
                  type: string
                  enum: [testnet, mainnet]
                  default: mainnet
                  description: Network this key is associated with
                rate_limit_tier:
                  type: string
                  enum: [free, standard, premium, internal]
                  default: standard
                  description: Rate limit tier for this key
                scope:
                  type: string
                  enum: [read, admin]
                  default: read
                  description: >
                    Permission scope for the new key. Distinct from the
                    ADMIN_API_KEY used to call this endpoint itself:
                    scope controls what the *new* database-issued key can do
                    on scope-gated data routes (see middleware.RequireScope).
                expires_at:
                  type: string
                  format: date-time
                  nullable: true
                  description: Optional expiry. Omit or null for a non-expiring key.
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      summary: List API keys
      description: List all API keys (admin only)
      operationId: listApiKeys
      tags:
        - API Keys
      security:
        - AdminKey: []
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/api-keys/{id}:
    patch:
      summary: Update an API key
      description: Update the label, rate limit tier, scope, or expiry of an active key (admin only)
      operationId: updateApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                rate_limit_tier:
                  type: string
                  enum: [free, standard, premium, internal]
                scope:
                  type: string
                  enum: [read, admin]
                expires_at:
                  type: string
                  description: >
                    RFC3339 timestamp to set a new expiry, or an empty string
                    to clear an existing expiry. Omit the field to leave it
                    unchanged.
      responses:
        "200":
          description: API key updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete an API key
      description: Revoke and delete an API key (admin only)
      operationId: deleteApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      responses:
        "204":
          description: API key deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/api-keys/{id}/rotate:
    post:
      summary: Rotate an API key
      description: >
        Creates a new key that inherits the old key's label, network, rate
        limit tier, and scope. The old key keeps authenticating for a grace
        window (default 24h) so callers can switch over, then is
        automatically revoked. Admin only.
      operationId: rotateApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID to rotate
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                grace_period_seconds:
                  type: integer
                  format: int64
                  description: Overrides the default rotation grace window, in seconds.
      responses:
        "201":
          description: New key created; old key enters its grace window
          content:
            application/json:
              schema:
                type: object
                properties:
                  rotated_from:
                    type: string
                    format: uuid
                  grace_until:
                    type: string
                    format: date-time
                  new_key:
                    $ref: "#/components/schemas/ApiKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: API key not found, already revoked, or already rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          description: The raw API key (only present in create/rotate responses, shown once)
        key_prefix:
          type: string
          description: First 16 characters of key for reference
        label:
          type: string
        network:
          type: string
        rate_limit_tier:
          type: string
        scope:
          type: string
          enum: [read, admin]
          description: Permission scope enforced by middleware.RequireScope on data routes
        created_by:
          type: string
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        request_count:
          type: integer
          format: int64
        revoked_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        grace_until:
          type: string
          format: date-time
          nullable: true
          description: Set by rotation; the old key authenticates until this time, then auto-revokes.
        rotated_from:
          type: string
          format: uuid
          nullable: true
        rotated_to:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code (e.g., INVALID_ARGUMENT, INTERNAL, UNAVAILABLE)
            message:
              type: string
              description: Human-readable error message
            request_id:
              type: string
              description: Request ID for debugging
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    AdminKey:
      type: apiKey
      in: header
      name: ADMIN_API_KEY
      description: Admin key for admin endpoints (environment variable)
...